Convert AddTaskRhoNew to static method
Conversion of AddTask macros to compile code allow easier usage
and greater compatibility with ROOT6.
The AddTaskRhoNew.C is kept for compatibility.
